+++ packages/entities/entities-plugins/src/components/free-form/shared/PluginFormWrapper.vue
@@ -21,7 +21,7 @@
 
     <div class="sc-form-config-form">
       <slot
-        :data="model"
+        :data="pruneData(model)"
         :schema="freeFormSchema"
         @change="onFormChange"
       />
@@ -29,11 +29,11 @@
   </KCollapse>
 </template>
 
-<script lang="ts" generic="T">
-export type Props<T> = {
+<script lang="ts">
+export type Props<T extends FreeFormPluginData> = {
   schema: FormSchema
   formSchema: any
-  model: Record<string, any>
+  model: T
   formModel: Record<string, any>
   formOptions: any
   isEditing: boolean
@@ -48,7 +48,8 @@ export type ConfigFormProps<T> = {
 }
 </script>
 
-<script setup lang="ts" generic="T">
+<script setup lang="ts" generic="T extends FreeFormPluginData">
+import { pick } from 'lodash-es'
 import { computed, ref, toValue, provide, type MaybeRefOrGetter } from 'vue'
 import { createI18n } from '@kong-ui-public/i18n'
 import { AUTOFILL_SLOT, AUTOFILL_SLOT_NAME } from '@kong-ui-public/forms'
@@ -56,6 +57,7 @@ import { VueFormGenerator } from '@kong-ui-public/forms'
 import { KCollapse } from '@kong/kongponents'
 import english from '../../../locales/en.json'
 import type { FormSchema } from '../../../types/plugins/form-schema'
+import type { FreeFormPluginData } from '../../../types/plugins/free-form'
 
 const { t } = createI18n<typeof english>('en-us', english)
 
@@ -115,6 +117,21 @@ const freeFormSchema = computed(() => {
 })
 
 const configCollapse = ref(false)
+
+/**
+ * Avoid passing freeform data that it can't handle. e.g. `scope`, `update_time`
+ * freeform will pass these unknown values back through the update method, resulting in the data being overwritten when it is eventually merged with the vfg's data
+ */
+function pruneData(data: Props<T>['model']) {
+  const ffDataKeys: (keyof Props<T>['model'])[] = [
+    'config',
+    'instance_name',
+    'partials',
+    'protocols',
+    'tags',
+  ]
+  return pick(data, ffDataKeys)
+}
 </script>
 
 <style lang="scss" scoped>