+++ packages/analytics/analytics-utilities/CHANGELOG.md
@@ -3,6 +3,17 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+# [9.5.0](https://github.com/Kong/public-ui-components/compare/@kong-ui-public/analytics-utilities@9.4.0...@kong-ui-public/analytics-utilities@9.5.0) (2025-05-22)
+
+
+### Features
+
+* add global filters ([#2140](https://github.com/Kong/public-ui-components/issues/2140)) ([0ddf722](https://github.com/Kong/public-ui-components/commit/0ddf7223ba2067024c4781b6cd53c906f89bf636))
+
+
+
+
+
 # [9.4.0](https://github.com/Kong/public-ui-components/compare/@kong-ui-public/analytics-utilities@9.3.4...@kong-ui-public/analytics-utilities@9.4.0) (2025-05-15)