+++ packages/analytics/dashboard-renderer/CHANGELOG.md
@@ -3,6 +3,41 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [5.4.1](https://github.com/Kong/public-ui-components/compare/@kong-ui-public/dashboard-renderer@5.4.0...@kong-ui-public/dashboard-renderer@5.4.1) (2025-05-22)
+
+**Note:** Version bump only for package @kong-ui-public/dashboard-renderer
+
+
+
+
+
+# [5.4.0](https://github.com/Kong/public-ui-components/compare/@kong-ui-public/dashboard-renderer@5.3.3...@kong-ui-public/dashboard-renderer@5.4.0) (2025-05-22)
+
+
+### Features
+
+* add global filters ([#2140](https://github.com/Kong/public-ui-components/issues/2140)) ([0ddf722](https://github.com/Kong/public-ui-components/commit/0ddf7223ba2067024c4781b6cd53c906f89bf636))
+
+
+
+
+
+## [5.3.3](https://github.com/Kong/public-ui-components/compare/@kong-ui-public/dashboard-renderer@5.3.2...@kong-ui-public/dashboard-renderer@5.3.3) (2025-05-21)
+
+**Note:** Version bump only for package @kong-ui-public/dashboard-renderer
+
+
+
+
+
+## [5.3.2](https://github.com/Kong/public-ui-components/compare/@kong-ui-public/dashboard-renderer@5.3.1...@kong-ui-public/dashboard-renderer@5.3.2) (2025-05-20)
+
+**Note:** Version bump only for package @kong-ui-public/dashboard-renderer
+
+
+
+
+
 ## [5.3.1](https://github.com/Kong/public-ui-components/compare/@kong-ui-public/dashboard-renderer@5.3.0...@kong-ui-public/dashboard-renderer@5.3.1) (2025-05-19)
 
 **Note:** Version bump only for package @kong-ui-public/dashboard-renderer