+++ packages/entities/entities-redis-configurations/CHANGELOG.md
@@ -3,6 +3,46 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [0.5.30](https://github.com/Kong/public-ui-components/compare/@kong-ui-public/entities-redis-configurations@0.5.29...@kong-ui-public/entities-redis-configurations@0.5.30) (2025-05-22)
+
+**Note:** Version bump only for package @kong-ui-public/entities-redis-configurations
+
+
+
+
+
+## [0.5.29](https://github.com/Kong/public-ui-components/compare/@kong-ui-public/entities-redis-configurations@0.5.28...@kong-ui-public/entities-redis-configurations@0.5.29) (2025-05-22)
+
+**Note:** Version bump only for package @kong-ui-public/entities-redis-configurations
+
+
+
+
+
+## [0.5.28](https://github.com/Kong/public-ui-components/compare/@kong-ui-public/entities-redis-configurations@0.5.27...@kong-ui-public/entities-redis-configurations@0.5.28) (2025-05-21)
+
+**Note:** Version bump only for package @kong-ui-public/entities-redis-configurations
+
+
+
+
+
+## [0.5.27](https://github.com/Kong/public-ui-components/compare/@kong-ui-public/entities-redis-configurations@0.5.26...@kong-ui-public/entities-redis-configurations@0.5.27) (2025-05-21)
+
+**Note:** Version bump only for package @kong-ui-public/entities-redis-configurations
+
+
+
+
+
+## [0.5.26](https://github.com/Kong/public-ui-components/compare/@kong-ui-public/entities-redis-configurations@0.5.25...@kong-ui-public/entities-redis-configurations@0.5.26) (2025-05-21)
+
+**Note:** Version bump only for package @kong-ui-public/entities-redis-configurations
+
+
+
+
+
 ## [0.5.25](https://github.com/Kong/public-ui-components/compare/@kong-ui-public/entities-redis-configurations@0.5.24...@kong-ui-public/entities-redis-configurations@0.5.25) (2025-05-20)
 
 **Note:** Version bump only for package @kong-ui-public/entities-redis-configurations