+++ packages/entities/entities-routes/CHANGELOG.md
@@ -3,6 +3,22 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [3.13.6](https://github.com/Kong/public-ui-components/compare/@kong-ui-public/entities-routes@3.13.5...@kong-ui-public/entities-routes@3.13.6) (2025-05-21)
+
+**Note:** Version bump only for package @kong-ui-public/entities-routes
+
+
+
+
+
+## [3.13.5](https://github.com/Kong/public-ui-components/compare/@kong-ui-public/entities-routes@3.13.4...@kong-ui-public/entities-routes@3.13.5) (2025-05-21)
+
+**Note:** Version bump only for package @kong-ui-public/entities-routes
+
+
+
+
+
 ## [3.13.4](https://github.com/Kong/public-ui-components/compare/@kong-ui-public/entities-routes@3.13.3...@kong-ui-public/entities-routes@3.13.4) (2025-05-19)
 
 **Note:** Version bump only for package @kong-ui-public/entities-routes