+++ packages/entities/entities-upstreams-targets/CHANGELOG.md
@@ -3,6 +3,14 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [3.7.31](https://github.com/Kong/public-ui-components/compare/@kong-ui-public/entities-upstreams-targets@3.7.30...@kong-ui-public/entities-upstreams-targets@3.7.31) (2025-05-21)
+
+**Note:** Version bump only for package @kong-ui-public/entities-upstreams-targets
+
+
+
+
+
 ## [3.7.30](https://github.com/Kong/public-ui-components/compare/@kong-ui-public/entities-upstreams-targets@3.7.29...@kong-ui-public/entities-upstreams-targets@3.7.30) (2025-05-19)
 
 **Note:** Version bump only for package @kong-ui-public/entities-upstreams-targets